+++ categories_sql_lite/lib/data/repositories/group_repository_impl.dart
@@ -1,35 +1,64 @@
 
 import 'package:categories_sql_lite/core/core.dart';
-import 'package:categories_sql_lite/data/db/database.dart';
+import 'package:categories_sql_lite/data/data.dart';
 import 'package:categories_sql_lite/domain/domain.dart';
 import 'package:sqflite/sqflite.dart'
 if(dart.library.io.Platform.isWindows)'package:sqflite_common_ffi/sqflite_ffi.dart';
 
-class GroupRepositoryImpl extends GroupRepository{
-  @override
-  Future<(int, int)> deleteGroup(int gid) async {
-    return await DBProvider.db.deleteGroup(gid);
+class GroupRepositoryImpl extends GroupsRepository{
+
+  GroupRepositoryImpl(){
+    groupsModel = GroupsModel();
   }
 
   @override
-  Future<List<Group>> getGroups() async {
-    return await DBProvider.db.getGroups();
+  Future<(int, int)> deleteGroup(Group value) async {
+    int? gid = value.gid;
+    try {
+      if (gid != null) {
+        groupsModel.group.remove(value);
+        return await DBProvider.db.deleteGroup(gid);
+      }
+    } catch(e,t){
+      Logger.print('Error $e\n$t', name: 'log', level: 0, error: false);
+      throw('Error deleteGroup: $e\n$t');
+    }
+    throw('Error deleteGroup');
   }
 
   @override
-  Future<Group> insertGroup(Group data,
+  Future<bool> getGroups() async {
+    try {
+      groupsModel.group.addAll(
+        await DBProvider.db.getGroups());
+      return groupsModel.group.isNotEmpty;
+    } catch(e,t){
+      Logger.print('Error $e\n$t', name: 'log', level: 0, error: false);
+      throw('Error getGroups: $e\n$t');
+    }
+  }
+
+  @override
+  Future<void> insertGroup(Group value,
       {
         ConflictAlgorithm conflictAlgorithm = ConflictAlgorithm.ignore
       }) async {
-    return await DBProvider.db.insertGroup(data, conflictAlgorithm:conflictAlgorithm);
+    try{
+      groupsModel.group.add(await DBProvider.db.insertGroup(value, conflictAlgorithm:conflictAlgorithm));
+    } catch(e,t){
+      Logger.print('Error $e\n$t', name: 'log', level: 0, error: false);
+      throw('Error insertGroup: $e\n$t');
+    }
   }
 
   @override
-  Future<int> updateGroup(Group data,
+  Future<int> updateGroup(Group oldValue, Group value,
       {
         ConflictAlgorithm conflictAlgorithm = ConflictAlgorithm.ignore
       }) async {
-    return await DBProvider.db.updateGroup(data, conflictAlgorithm: conflictAlgorithm);
+    groupsModel.group.remove(oldValue);
+    groupsModel.group.add(value);
+    return await DBProvider.db.updateGroup(value, conflictAlgorithm: conflictAlgorithm);
   }
 
 }
